[http-client-java]mgmt, fix instant max value (#9123)
Root cause:
`return RANDOM.nextInt() & Long.MAX_VALUE;`

`RANDOM.nextInt()` sometimes will return negative integer. 
With `& Long.MAX_VALUE`, it'll be promoted to long, with negative sign
removed.
The promoted number will exceed Integer.MAX_VALUE.

- bug fix in 65fa31c5eab7e6ecac1028eb4dee9e3b2d39402f
- (improvement)Group primitive logic in `WireTypeClientTypeConverter`.
- (improvement)Added test case for primitive types.
- regen & test validation in
https://github.com/Azure/autorest.java/pull/3233

+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/util/ModelTestCaseUtil.java
@@ -110,7 +110,7 @@ public class ModelTestCaseUtil {
             return randomString();
         } else if (type.asNullable() == ClassType.UNIX_TIME_LONG) {
             // use nextInt to avoid exceeding unixTime limit
-            return RANDOM.nextInt() & Long.MAX_VALUE;
+            return RANDOM.nextInt() & Integer.MAX_VALUE;
         } else if (type == ClassType.DATE_TIME) {
             return randomDateTime().toString();
         } else if (type == ClassType.DATE_TIME_RFC_1123) {
